[java] Remove usage of JavaTokenKinds...
outside of AST and CPD packages
